+++ assets/svelte/src/components/CoursePicker.svelte
@@ -1,9 +1,11 @@
 <script>
-    let { courses } = $props();
+    import CourseSelector from "./CourseSelector.svelte";
+
+    let { courses, selectedCourses = $bindable() } = $props();
 </script>
 
-<div class="course-list">
+<div class="course-list bordered col">
     {#each courses as course}
-        {course.title}
+        <CourseSelector {course} bind:selected={selectedCourses[course.id]} />
     {/each}
 </div>